requirements.txt
fastapi
uvicorn
pymupdf
pymupdf4llm
pdfplumber
sentence-transformers
python-multipart
sqlalchemy
psycopg2-binary
pgvector
alembic
python-dotenv
pillow
pytesseract
langgraph
pytest
langchain-text-splitters
tiktoken
